refactor: 일정 타입 네이밍 변경
